Refactor category and product models: introduce slug field, update related methods, and enhance validation

+++ src/main/java/hoanght/posapi/service/impl/CategoryServiceImpl.java
@@ -7,13 +7,10 @@ import hoanght.posapi.exception.NotFoundException;
 import hoanght.posapi.model.Category;
 import hoanght.posapi.repository.jpa.CategoryRepository;
 import hoanght.posapi.service.CategoryService;
+import hoanght.posapi.utils.StringUtils;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
-import org.springframework.cache.annotation.CacheEvict;
-import org.springframework.cache.annotation.CachePut;
-import org.springframework.cache.annotation.Cacheable;
-import org.springframework.cache.annotation.Caching;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -27,42 +24,37 @@ public class CategoryServiceImpl implements CategoryService {
     private final ModelMapper modelMapper;
 
     @Override
-    @Cacheable(value = "category", key = "#categoryId")
     public Category findById(Long categoryId) {
-        return categoryRepository.findById(categoryId)
-                .orElseThrow(() -> new NotFoundException("Category with ID " + categoryId + " not found"));
+        return categoryRepository.findById(categoryId).orElseThrow(() -> new NotFoundException("Category with ID " + categoryId + " not found"));
     }
 
     @Override
-    @Cacheable(value = "categories")
     public Page<Category> findAll(Pageable pageable) {
         return categoryRepository.findAll(pageable);
     }
 
     @Override
     @Transactional
-    @CacheEvict(value = "categories", allEntries = true)
     public Category create(CategoryCreationRequest categoryCreationRequest) {
         if (categoryRepository.existsByName(categoryCreationRequest.getName())) {
             throw new AlreadyExistsException("Category with name " + categoryCreationRequest.getName() + " already exists");
         }
         Category category = modelMapper.map(categoryCreationRequest, Category.class);
+        category.setSlug(generateSlug(category.getName()));
         return categoryRepository.save(category);
     }
 
     @Override
     @Transactional
-    @CachePut(value = "category", key = "#categoryId")
-    @CacheEvict(value = "categories", allEntries = true)
     public Category update(Long categoryId, CategoryUpdateRequest categoryUpdateRequest) {
-        Category existingCategory = categoryRepository.findById(categoryId)
-                .orElseThrow(() -> new NotFoundException("Category with ID " + categoryId + " not found"));
+        Category existingCategory = categoryRepository.findById(categoryId).orElseThrow(() -> new NotFoundException("Category with ID " + categoryId + " not found"));
 
         if (categoryUpdateRequest.getName() != null) {
-            if (categoryRepository.existsByName(categoryUpdateRequest.getName()) && !existingCategory.getName().equals(categoryUpdateRequest.getName())) {
+            if (categoryRepository.existsByNameAndIdNot(categoryUpdateRequest.getName(), existingCategory.getId()) && !existingCategory.getName().equals(categoryUpdateRequest.getName())) {
                 throw new AlreadyExistsException("Category with name " + categoryUpdateRequest.getName() + " already exists");
             }
             existingCategory.setName(categoryUpdateRequest.getName());
+            existingCategory.setSlug(generateSlug(existingCategory.getName()));
         }
         Optional.ofNullable(categoryUpdateRequest.getImageUrl()).ifPresent(existingCategory::setImageUrl);
         return categoryRepository.save(existingCategory);
@@ -70,14 +62,21 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     @Transactional
-    @Caching(evict = {
-            @CacheEvict(value = "category", key = "#categoryId"),
-            @CacheEvict(value = "categories", allEntries = true)
-    })
     public void delete(Long categoryId) {
         if (!categoryRepository.existsById(categoryId)) {
             throw new NotFoundException("Category with ID " + categoryId + " not found");
         }
+        if (categoryRepository.existsByIdAndProductsIsNotEmpty(categoryId))
+            throw new AlreadyExistsException("Cannot delete category with associated products");
         categoryRepository.deleteById(categoryId);
+    }
+
+    private String generateSlug(String name) {
+        int counter = 0;
+        String baseSlug = StringUtils.toSlug(name);
+        while (categoryRepository.existsBySlug(counter == 0 ? baseSlug : baseSlug + "-" + counter)) {
+            counter++;
+        }
+        return counter == 0 ? baseSlug : baseSlug + "-" + counter;
     }
 }
